(PC-13519)[PRO] test: fix tests
